Factor Worker class into its own file
